respond to PR review
